v0.6.1: critical bugfixes + settings module
- Fix secrets_get returns password value not key name
- Fix PTY race condition (await listeners before PTY spawn)
- Fix PTY listener leak (Map routing instead of shared array)
- Wire OSC 7 handler to update tab cwd from PTY data
- Fix CSP to allow preview iframes (frame-src)
- Fix Preview onUrlChange not passed to component
- New Settings module (theme, font size, AI provider config)
- Fix 5 pre-existing Rust clippy warnings

+++ src-tauri/src/modules/pty/mod.rs
@@ -158,8 +158,8 @@ pub fn pty_create(
 
     let pair = pty_system
         .openpty(PtySize {
-            rows: args.rows as u16,
-            cols: args.cols as u16,
+            rows: args.rows,
+            cols: args.cols,
             pixel_width: 0,
             pixel_height: 0,
         })
